ITraverser: Refactor out traverser into a support module
