+++ src/Snowflake.Framework.Tests/Snowflake.Framework.Tests.csproj
@@ -41,6 +41,7 @@
     <ProjectReference Include="..\Snowflake.Support.PluginManager\Snowflake.Support.PluginManager.csproj" />
     <ProjectReference Include="..\Snowflake.Support.Remoting.GraphQl.Framework\Snowflake.Support.Remoting.GraphQl.Framework.csproj" />
     <ProjectReference Include="..\Snowflake.Support.Remoting.GraphQl.RootProvider\Snowflake.Support.Remoting.GraphQl.RootProvider.csproj" />
+    <ProjectReference Include="..\Snowflake.Support.Scraping.RecordTraversers\Snowflake.Support.Scraping.RecordTraversers.csproj" />
     <ProjectReference Include="..\Snowflake.Support.ShiragameProvider\Snowflake.Support.ShiragameProvider.csproj" />
     <ProjectReference Include="..\Snowflake.Support.StoneProvider\Snowflake.Support.StoneProvider.csproj" />
   </ItemGroup>
