+++ src/Snowflake.Framework.Utility/EnumerableExtensions.cs
@@ -23,11 +23,11 @@ namespace Snowflake.Utility
         /// <summary>
         /// Returns values distinct by a predicate.
         /// </summary>
-        /// <typeparam name="TSource">The source type</typeparam>
-        /// <typeparam name="TKey">The predicate to return/typeparam>
+        /// <returns>The distinct items.</returns>
         /// <param name="source">The enumerable to return.</param>
         /// <param name="keySelector">The predicate to select on.</param>
-        /// <returns>The distinct items.</returns>
+        /// <typeparam name="TSource">The source type</typeparam>
+        /// <typeparam name="TKey">The predicate to return/typeparam>
         public static IEnumerable<TSource> DistinctBy<TSource, TKey>(this IEnumerable<TSource> source, Func<TSource, TKey> keySelector)
         {
             HashSet<TKey> seenKeys = new HashSet<TKey>();
