--- src/Snowflake.Framework/Scraping/Traverser.cs
@@ -1,10 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace Snowflake.Scraping
-{
-    class Traverser
-    {
-    }
-}
