--- src/Snowflake.Support.ScrapeProvider/ScrapeEngine.cs
@@ -1,10 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace Snowflake.Support.ScrapeProvider
-{
-    class ScrapeEngine
-    {
-    }
-}
